unifying problem classes; adding algorithm classes
