+++ pyrlap/pyrlap2/core/__init__.py
@@ -1,19 +1,6 @@
 from pyrlap.pyrlap2.core.distributions import Multinomial, Distribution
-from pyrlap.pyrlap2.core.mdp.mdp import \
-    MarkovDecisionProcess, \
-    ANDMarkovDecisionProcess
-from pyrlap.pyrlap2.core.mdp.tabularmdp import \
-    TabularMarkovDecisionProcess, \
-    ANDTabularMarkovDecisionProcess
-from pyrlap.pyrlap2.core.mdp.policy import Policy, TabularPolicy
 
-from pyrlap.pyrlap2.core.posg.posg import \
-    PartiallyObservableStochasticGame, \
-    ANDPartiallyObservableStochasticGame 
-
-from pyrlap.pyrlap2.core.stochasticgame import \
-	StochasticGame, ANDStochasticGame
-
-from pyrlap.pyrlap2.core.assignment.assignmentmap import AssignmentMap, DefaultAssignmentMap
-from pyrlap.pyrlap2.core.assignment.assignmentset import AssignmentSet
-from pyrlap.pyrlap2.core.assignment.assignmentcache import AssignmentCache
+from pyrlap.pyrlap2.core.problemclasses.mdp import *
+from pyrlap.pyrlap2.core.problemclasses.mdp.policy import *
+from pyrlap.pyrlap2.core.problemclasses.stochasticgame import *
+from pyrlap.pyrlap2.core.assignment import *
