+++ pyrlap/pyrlap2/core/assignment/__init__.py
@@ -1,0 +1,7 @@
+from pyrlap.pyrlap2.core.assignment.assignmentmap import \
+    AssignmentMap, \
+    DefaultAssignmentMap
+from pyrlap.pyrlap2.core.assignment.assignmentset import \
+    AssignmentSet
+from pyrlap.pyrlap2.core.assignment.assignmentcache import \
+    AssignmentCache
