--- pyrlap/pyrlap2/core/mdp/policy/__init__.py
@@ -1,2 +1,0 @@
-from pyrlap.pyrlap2.core.mdp.policy.policy import Policy
-from pyrlap.pyrlap2.core.mdp.policy.tabularpolicy import TabularPolicy
