+++ pyrlap/pyrlap2/core/problemclasses/mdp/__init__.py
@@ -1,9 +1,9 @@
-from pyrlap.pyrlap2.core.mdp.mdp import \
+from pyrlap.pyrlap2.core.problemclasses.mdp.mdp import \
     MarkovDecisionProcess, \
     ANDMarkovDecisionProcess
-from pyrlap.pyrlap2.core.mdp.tabularmdp import \
+from pyrlap.pyrlap2.core.problemclasses.mdp.tabularmdp import \
     TabularMarkovDecisionProcess, \
     ANDTabularMarkovDecisionProcess
-from pyrlap.pyrlap2.core.mdp.factoredmdp import \
+from pyrlap.pyrlap2.core.problemclasses.mdp.factoredmdp import \
     FactoredMarkovDecisionProcess, \
     ANDFactoredMarkovDecisionProcess
