+++ pyrlap/pyrlap2/core/problemclasses/mdp/policy/__init__.py
@@ -1,0 +1,2 @@
+from pyrlap.pyrlap2.core.problemclasses.mdp.policy.tabularpolicy import TabularPolicy
+from pyrlap.pyrlap2.core.problemclasses.mdp.policy.policy import Policy
