+++ pyrlap/pyrlap2/core/problemclasses/problemclass.py
@@ -1,0 +1,5 @@
+from abc import ABC, abstractmethod
+
+class ProblemClass(ABC):
+    """Abstract superclass for all problem classes"""
+    pass
