+++ pyrlap/pyrlap2/core/problemclasses/stochasticgame/__init__.py
@@ -1,0 +1,2 @@
+from pyrlap.pyrlap2.core.problemclasses.stochasticgame.stochasticgame \
+    import StochasticGame, ANDStochasticGame
