--- pyrlap/pyrlap2/core/stochasticgame/__init__.py
@@ -1,2 +1,0 @@
-from pyrlap.pyrlap2.core.stochasticgame.stochasticgame \
-    import StochasticGame, ANDStochasticGame
