+++ pyrlap/pyrlap2/tests/domaintests.py
@@ -1,6 +1,7 @@
 import unittest
 import numpy as np
 from pyrlap.pyrlap2.domains import GridWorld
+from pyrlap.pyrlap2.domains.gridgame.gridgame import GridGame
 from pyrlap.pyrlap2.domains import StickyActionMDP
 from pyrlap.pyrlap2.algorithms import VectorizedValueIteration
 
@@ -33,8 +34,23 @@ class DomainTestCase(unittest.TestCase):
         vi.planOn(sagw)
         saTraj = vi.policy.runOn(sagw)['stateTraj']
         self.assertTrue(saTraj[0] == {'curAction': {'dx': 0, 'dy': 0}, 'groundState': {'x': 8, 'y': 0}})
+
+    def test_gridgame_initialization(self):
+        #example usage
+        gameString = """
+            #  # # #  G0 #  # # # 
+            G0 . . A0 .  A1 . . G1
+            #  # # #  G1 #  # # #
+        """.strip()
+
+        gg = GridGame(gameString)
+        s = gg.getInitialStateDist().sample()
+        a = {'A0': {'x': 1, 'y': 0}, 'A1': {'x': -1, 'y': 0}}
+        nsdist = gg.getNextStateDist(s, a)
+        ns = nsdist.sample()
+        r = gg.getJointRewards(s, a, ns)
+        self.assertTrue(True)
         
 if __name__ == '__main__':
     unittest.main()
-
